fix: dark button custom styles
