+++ docs/examples/button/custom.vue
@@ -1,17 +1,15 @@
+<script lang="ts" setup>
+import { isDark } from '~/composables/dark'
+</script>
+
 <template>
   <div class="flex">
-    <el-button color="#626aef">Light</el-button>
-    <el-button color="#626aef" plain>Light</el-button>
+    <el-button color="#626aef" :dark="isDark">Light</el-button>
+    <el-button color="#626aef" :dark="isDark" plain>Light</el-button>
 
-    <el-button color="#626aef" dark>Dark</el-button>
-    <el-button color="#626aef" dark plain>Dark</el-button>
-  </div>
-
-  <div class="flex" m="t-4">
-    <el-button color="#626aef" disabled>Disabled</el-button>
-    <el-button color="#626aef" disabled plain>Disabled</el-button>
-
-    <el-button color="#626aef" disabled dark>Disabled</el-button>
-    <el-button color="#626aef" disabled dark plain>Disabled</el-button>
+    <el-button color="#626aef" :dark="isDark" disabled>Disabled</el-button>
+    <el-button color="#626aef" :dark="isDark" disabled plain
+      >Disabled</el-button
+    >
   </div>
 </template>
