+++ packages/theme-chalk/src/mixins/_button.scss
@@ -108,8 +108,8 @@
     ),
     'disabled': (
       'text-color': (
-        'disabled',
-        'text-color',
+        'color',
+        'white',
       ),
       'bg-color': (
         'color',
